Modify trim clips method
